requirements.txt
aiohttp
asyncio
discord-py-interactions
discord.py
python-dotenv